refactor: unify into api for command analyzer
